# Conflicts:
#	Project_DR/Assets/01_DR/Fonts_TMP/GmarketSansTTFMedium SDF for Inventory.asset
#	Project_DR/Assets/01_DR/Fonts_TMP/GmarketSansTTFMedium SDF for Shop.asset
#	Project_DR/Assets/01_DR/ImportAssets/PolygonParticleFX/Materials/PolygonParticleFX_Rock_01.mat
#	Project_DR/Assets/01_DR/Scripts/4_SG/01.Prefabs/NPC.meta
#	Project_DR/Assets/01_DR/Scripts/4_SG/01.Prefabs/NPC/NPCCanvasPrefab.prefab
#	Project_DR/Assets/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Fallback.asset
#	Project_DR/ProjectSettings/EditorBuildSettings.asset

+++ Project_DR/Assets/01_DR/Scripts/3_JS/DB/PlayerData.cs
@@ -3,16 +3,16 @@ using System;
 [Serializable]
 public class PlayerData
 {
-    public float hp;
+    public int hp;
     public int gold;
-    public float exp;
-    public float gold_increase;
-    public float exp_increase;
-    public float weapon_atk;
-    public float weapon_cri_rate;
-    public float weapon_cri_damage;
-    public float weapon_atk_rate;
-    public float weapon_exp;
+    public int exp;
+    public int gold_increase;
+    public int exp_increase;
+    public int weapon_atk;
+    public int weapon_cri_rate;
+    public int weapon_cri_damage;
+    public int weapon_atk_rate;
+    public int weapon_exp;
     public int skill_level_1;
     public int skill_level_2;
     public int skill_level_3;
